+++ client/content/blog/posts/2021/04/17/index.js
@@ -1,0 +1,9 @@
+const React = require('react');
+
+const content = (
+  <div>
+    <h1>So you wanna make a blog...</h1>
+  </div>
+);
+
+module.exports = content;